Add files via upload
This upload includes a css file which creates a full-width view. It has a header with title/subtitle, background color, inline menu positioned at the bottom of the header, which includes link states, and active page id assigned to correct list item. It includes a percentage margin on main content, and header content, as well as some styling to the paragraph and heading levels. Plus it includes a styled block quote with its own paragraph styles to create an indented section.